backend/requirements.txt
fastapi==0.135.1
groq==1.1.1
openai==2.29.0
pydantic==2.12.5
pytest==9.0.2
python-dotenv==1.2.2
supabase==2.28.2
thefuzz==0.22.1
uvicorn==0.23.2